feature(read-entry): complete implementation for readEntry
- remove entryList container
- move getEntry implementation to home container
- change currentEntry to use array index instead of entry id
- return descriptive error on getEntry and createEntry
- complete implementation

finishes #162298220
